+++ src/components/routes/RoutesPicsList.js
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from 'react'
 import { RoutesContext } from './RouteProvider'
 import { useParams } from 'react-router'
 import { RoutesPicsCard } from './RoutesPicsCard'
+// import ImageGallery from 'react-image-gallery'
 import "./Routes.css"
 
 
@@ -13,6 +14,9 @@ export const RoutePicsList = () => {
 
 	const {routeId} = useParams()
 
+	// const images = []
+
+
 	useEffect(()=>{
 		getRoutePics()
 	},[])
@@ -21,14 +25,27 @@ export const RoutePicsList = () => {
 		const filtered = routePics.filter(rp => rp.routeId === parseInt(routeId))
 		console.log(filtered)
 		setFilteredPics(filtered)
+		// filtered.map(pic => {
+		// 	const item = {
+		// 		original: pic.URL,
+    		// 		thumbnail: 'https://picsum.photos/id/1018/250/150/',
+		// 	}
+		// 	return images.push(item)})
+			
 	},[routePics, routeId])
+
+	
 
 
 	return(
 		<>
+
+		{/* <ImageGallery items={images} /> */}
 		{filteredPics.map(routePic => {
+			
 			return <RoutesPicsCard key={routePic.id} routePic={routePic} />
 		})}
+		
 		</>
 	)
 }
